renaming stretch->real world and floorplanner->hssd
