+++ README.md
@@ -136,14 +136,14 @@ python src/home_robot_hw/home_robot_hw/nodes/simple_grasp_server.py
 Then you can run the OVMM example script:
 ```
 cd $HOME_ROBOT_ROOT
-python projects/stretch_ovmm/eval_episode.py
+python projects/real_world_ovmm/eval_episode.py
 ```
 
 #### 7. Simulation Setup
 
 To set up the simulation stack with Habitat, train DDPPO skills and run evaluations: see the [installation instructions](src/home_robot_sim/README.md) in `home_robot_sim`.
 
-For more details on the OVMM challenge, see the [Habitat OVMM readme](projects/stretch_ovmm/README.md).
+For more details on the OVMM challenge, see the [Habitat OVMM readme](projects/real_world_ovmm/README.md).
 
 
 ## Code Contribution
