+++ projects/habitat_ovmm/README.md
@@ -30,7 +30,7 @@ git clone https://huggingface.co/datasets/osmm/episodes
 
 Run
 ```
-python projects/habitat_ovmm/eval_vectorized.py
+python projects/habitat_ovmm/eval_dataset.py
 ```
 
 Results are saved to `datadump/images/eval_floorplanner/`.
@@ -56,18 +56,18 @@ cd -
 
 ## Run
 
-> Note: Ensure `GROUND_TRUTH_SEMANTICS:0` in `configs/agent/floorplanner_eval.yaml` to test DETIC perception.
+> Note: Ensure `GROUND_TRUTH_SEMANTICS:0` in `configs/agent/hssd_eval.yaml` to test DETIC perception.
 
 ```
 cd /path/to/home-robot
 
 
 # Evaluation on complete episode dataset with GT semantics
-python projects/habitat_ovmm/eval_vectorized.py
+python projects/habitat_ovmm/eval_dataset.py
 
 # Evaluation on complete episode dataset with DETIC
-python projects/habitat_ovmm/eval_vectorized.py  --baseline_config_path projects/habitat_ovmm/configs/agent/floorplanner_detic_eval.yaml
+python projects/habitat_ovmm/eval_dataset.py  --baseline_config_path projects/habitat_ovmm/configs/agent/floorplanner_detic_eval.yaml
 
 # Evaluation on specific episodes
-python projects/habitat_ovmm/eval_vectorized.py habitat.dataset.episode_ids="[151,182]"
+python projects/habitat_ovmm/eval_dataset.py habitat.dataset.episode_ids="[151,182]"
 ```
