+++ projects/real_world_ovmm/configs/example_cat_map.json
