+++ projects/real_world_ovmm/eval_episode.py
@@ -21,7 +21,7 @@ from home_robot_hw.utils.config import load_config
 @click.option("--start-recep", default="table")
 @click.option("--goal-recep", default="chair")
 @click.option(
-    "--cat-map-file", default="projects/stretch_ovmm/configs/example_cat_map.json"
+    "--cat-map-file", default="projects/real_world_ovmm/configs/example_cat_map.json"
 )
 @click.option("--max-num-steps", default=200)
 @click.option("--visualize-maps", default=False, is_flag=True)
