+++ projects/real_world_ovmm/pick_cup_real_robot.py
