+++ projects/real_world_ovmm/utils/rotate_left.py
