+++ src/home_robot_hw/home_robot_hw/utils/config.py
@@ -7,7 +7,7 @@ def load_config(
     """Load config path for real world experiments and use proper presets."""
     if config_path is None:
         # TODO: make sure this is the right default
-        config_path = "projects/stretch_ovmm/configs/agent/floorplanner_eval.yaml"
+        config_path = "projects/real_world_ovmm/configs/agent/eval.yaml"
     config, config_str = get_config(config_path)
     config.defrost()
     config.NUM_ENVIRONMENTS = 1
